Modifies the CNN model training script
This commit modifies the CNN model training script. It adds a function modifying the loaded model for transfer learning without and with fine-tuning step. It also extend the training function so the training can be started. In the further work the logging functionality and the plotting functionality of training results will be added. After initial training runs, one epoch would take approximately 20 minutes. The whole training would take 5 hours. After the augmentation of the dataset to 4 times its size, the time would not be feasible on my computer. Possibility to use Google Colab will be explored.

Implemented by: Jakub Dunaj
Refs.: Issue #2