Create a script limiting the number of dataset images in a class to min(true_number, LIMIT) images
This commit adds as script  limiting the number of dataset images in a class to min(true_number, LIMIT) images. The other modules in the preprocessing pipeline had to be adjusted too. New pipeline had been set up in the dataset_preprocessing_wrapper.py module, so that the whole preprocessing pipeline can be executed easily. This commit also includes a picture_counter.py utility module to count number of images in a dataset.

Implemented by: Jakub Dunaj
Refs.: Issue #17