Extends the accuracy_per_class_metric.py module
This commit extends the accuracy_per_class_metric.py module. The challenge is to come up with a way how to update the variables when the tensors are symbolic. This means that the values in the tensors cannot be accessed directly. The keras.ops package gives many operations to works with symbolic tensors.

Implemented by: Jakub Dunaj
Refs.: Issue #20